Make debug with eric6 work on Windows10
While testing eric6 (installed via pip) on a Windows 10 machine,
I got exception that sys.stdin was None (and mode was not an attribute
of it). So after some googling I found that sometimes when Python
is started from a gui on Windows, that sys.stdin/stderr/stdout are
pointing to invalid file descriptors. According to the link in the code
you can assign these to them.
And then the plugin also works on Windows 10 with me here!